imdb_recommendation/requirements.txt
IMDbPY==2022.7.9
